# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HookLab platform core API",
    "version": "0.3.0-rc.5",
    "description": "Core PostgreSQL platform routes. Experimental contract, not a complete listing of administrative routes. All secrets and event bodies must be handled as private data."
  },
  "servers": [{"url": "http://127.0.0.1:8787", "description": "Loopback development only"}],
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}},
...
  },
  "paths": {
...
    "/api/admin/tenants": {"post": {"summary": "Create tenant with bootstrap token", "security": [{"bearerAuth": []}],
      "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["id", "name"], "properties": {"id": {"type": "string"}, "name": {"type": "string"}}}}}},
      "responses": {"201": {"description": "Tenant and one-time owner token"}, "401": {"description": "Invalid bootstrap token"}}}},
...
  }
}